CHE-84 Image controller directory restructure
