+++ server/controllers/imageController/index.ts
@@ -1,0 +1,4 @@
+import uploadProfilePicture from './uploadProfilePicture/uploadProfilePicture';
+import generatePresignedUrl from './generatePresignedUrl/generatePresignedUrl';
+
+export { uploadProfilePicture, generatePresignedUrl };
